+++ src/reader/app/pages.ts
@@ -2,4 +2,5 @@ export {
     isYomuHostedAppUrl,
     isYomuHostedPassivePage,
     isYomuHostedVideoPlayerPage,
+    isYomuHostedPdfReaderPage,
 } from './pages-url';